cashprint_project/requirements.txt
asgiref==3.8.1
Django==5.2.3
python-decouple==3.8
sqlparse==0.5.3

Pillow==10.0.1
django-crispy-forms==2.0
crispy-bootstrap4==2022.1
django-widget-tweaks==1.4.12
python-decouple==3.8
django-extensions==3.2.3
reportlab==4.0.4
django-phonenumber-field==7.1.0
phonenumbers==8.13.22
